initial organization and cleaunup from crete-react-app
