+++ client/src/App.js
@@ -1,0 +1,5 @@
+function App() {
+  return <div className="App">Hello World!!!</div>;
+}
+
+export default App;
